维护后端功能, 为mybatis添加分页插件功能

+++ PropSmart/backend/src/main/java/com/mrdotxin/propsmart/model/dto/facility/reservation/FacilityReservationQueryRequest.java
@@ -20,7 +20,7 @@ public class FacilityReservationQueryRequest extends PageRequest implements Seri
      * 设施ID
      */
     @Schema(description = "设施ID")
-    private Integer facilityId;
+    private Long facilityId;
     
     /**
      * 用户ID
