+++ PropSmart/backend/src/main/java/com/mrdotxin/propsmart/model/entity/FacilityReservation.java
@@ -30,7 +30,7 @@ public class FacilityReservation implements Serializable {
      * 设施ID
      */
     @TableField(value = "facilityId")
-    private Integer facilityId;
+    private Long facilityId;
 
     /**
      * 预订时间
